(feature) add the setting FTP view
This is just the view this do nothing.
For the next step we have to put the textfield to a private argument.
Set a prefered size for each of this field
See if we can make the asword field
add button (ConnectionTest/Validate)
and add the action perfomed

This is what we can do for terminate the view